init Mesh
