+++ Mesh.cpp
@@ -1,0 +1,15 @@
+# include <glm/glm.hpp>
+#include "Mesh.hpp"
+#include <iostream>
+
+void Mesh::draw_self(VisualServer* server) const {
+    std::cout << "Draw Mesh\n";
+    // std::vector<glm::vec4> triangle;
+    // // glm::vec4 vec = glm::vec4(0.0, 1.0, 0.0, 1.0);
+    // triangle.emplace_back(0.0, 0.0, 1.0, 1.0);
+    // triangle.emplace_back(1.0, 0.0, 0.0, 1.0);
+    // triangle.emplace_back(0.0, 1.0, 0.0, 1.0);
+    // server->draw_polygon(triangle);
+    
+}
+
