+++ Mesh.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "TreeObject.hpp"
+
+class Mesh: public TreeObject {
+	void draw_self(VisualServer* server) const override;
+};
